Move files, cleanup some things
