+++ project/algorithms/bases/algorithm_test.py
@@ -20,8 +20,8 @@ from typing_extensions import ParamSpec
 
 from project.configs.config import Config, cs
 from project.conftest import setup_hydra_for_tests_and_compose
+from project.datamodules.bases.image_classification import ImageClassificationDataModule
 from project.datamodules.bases.vision import VisionDataModule
-from project.datamodules.image_classification import ImageClassificationDataModule
 from project.experiment import (
     instantiate_datamodule,
     instantiate_network,
