+++ project/algorithms/bases/image_classification_test.py
@@ -14,9 +14,9 @@ from project.algorithms.bases.algorithm_test import (
     CheckBatchesAreTheSameAtEachStep,
 )
 from project.algorithms.bases.image_classification import ImageClassificationAlgorithm
-from project.datamodules.image_classification import ImageClassificationDataModule
-from project.datamodules.moving_mnist import MovingMnistDataModule
+from project.datamodules.bases.image_classification import ImageClassificationDataModule
 from project.datamodules.rl.rl_datamodule import RlDataModule
+from project.datamodules.vision.moving_mnist import MovingMnistDataModule
 from project.utils.types import DataModule
 
 ImageAlgorithmType = TypeVar("ImageAlgorithmType", bound=ImageClassificationAlgorithm)
