+++ project/configs/datamodule/__init__.py
@@ -16,11 +16,11 @@ from project.datamodules import (
     RlDataModule,
     VisionDataModule,
 )
-from project.datamodules.cifar10 import cifar10_train_transforms
-from project.datamodules.imagenet32 import imagenet32_train_transforms
-from project.datamodules.inaturalist import INaturalistDataModule, TargetType, Version
-from project.datamodules.mnist import mnist_train_transforms
-from project.datamodules.moving_mnist import MovingMnistDataModule
+from project.datamodules.vision.cifar10 import cifar10_train_transforms
+from project.datamodules.vision.imagenet32 import imagenet32_train_transforms
+from project.datamodules.vision.inaturalist import INaturalistDataModule, TargetType, Version
+from project.datamodules.vision.mnist import mnist_train_transforms
+from project.datamodules.vision.moving_mnist import MovingMnistDataModule
 
 FILE = Path(__file__)
 REPO_ROOTDIR = FILE.parent
