+++ project/networks/__init__.py
@@ -13,14 +13,14 @@ from .fcnet import FcNet
 class ResNet18Config:
     pretrained: bool = False
     num_classes: int = interpolated_field(
-        "${instance_attr:datamodule.num_classes,datamodule.action_dims}", 1000
+        "${instance_attr:datamodule.num_classes,datamodule.action_dims}", default=1000
     )
 
 
 @hydrated_dataclass(target=FcNet, hydra_recursive=True, hydra_convert="object")
 class FcNetConfig:
     output_dims: int = interpolated_field(
-        "${instance_attr:datamodule.num_classes,datamodule.action_dims}", 1
+        "${instance_attr:datamodule.num_classes,datamodule.action_dims}", default=1
     )
     hparams: FcNet.HParams = field(default_factory=FcNet.HParams)
 
@@ -37,10 +37,13 @@ class FcNetConfig:
 # - handler for creating the network from an RLDataModule
 # - ...
 
+# Currently, we're using something like option 1, where we use `interpolated_field` to retrieve
+# some attributes from the datamodule when creating the network configs.
+
 _cs = ConfigStore.instance()
 _cs.store(group="network", name="fcnet", node=FcNetConfig)
-# _cs.store(group="network", name="fcnet", node=FcNet.HParams)
 _cs.store(group="network", name="resnet18", node=ResNet18Config)
-
+...
+# Add your network configs here.
 
 __all__ = ["FcNet", "FcNetConfig", "ResNet18Config"]
