+++ project/networks/fcnet.py
@@ -8,7 +8,7 @@ import torch
 from gym.spaces.utils import flatdim
 from torch import Tensor, nn
 
-from project.datamodules.image_classification import ImageClassificationDataModule
+from project.datamodules.bases.image_classification import ImageClassificationDataModule
 from project.datamodules.rl.rl_datamodule import RlDataModule
 from project.utils.types.protocols import DataModule
